+++ src/pages/payments/constant.ts
@@ -1,0 +1,4 @@
+export const BRAND_NUMBER = {
+  visa: '4',
+  mastercard: ['51', '52', '53', '54', '55'],
+};
