+++ src/pages/payments/index.ts
@@ -1,1 +1,3 @@
 export { Payments } from './Payments';
+export type { ExpirationDate } from './types';
+export { BRAND_NUMBER } from './constant';
