+++ src/pages/payments/types.ts
@@ -1,0 +1,4 @@
+export type ExpirationDate = {
+  year: string;
+  month: string;
+};
